PUB-1260 SSCS Daily List template (#460)
* New list added for SSCS daily list

* Review comment

* Fixed output
