+++ src/main/modules/nunjucks/listTypeLookup.json
@@ -8,5 +8,6 @@
   "MAGS_STANDARD_LIST": "Magistrate Standard List",
   "CIVIL_DAILY_CAUSE_LIST": "Civil Daily Cause List",
   "FAMILY_DAILY_CAUSE_LIST": "Family Daily Cause List",
-  "CIVIL_AND_FAMILY_DAILY_CAUSE_LIST": "Civil And Family Daily Cause List"
+  "CIVIL_AND_FAMILY_DAILY_CAUSE_LIST": "Civil And Family Daily Cause List",
+  "SSCS_DAILY_LIST": "Social Security and Child Support Tribunal Daily List"
 }
