+++ src/main/modules/nunjucks/listUrlLookup.json
@@ -8,5 +8,6 @@
   "MAGS_STANDARD_LIST": "",
   "CIVIL_DAILY_CAUSE_LIST": "daily-cause-list",
   "FAMILY_DAILY_CAUSE_LIST": "family-daily-cause-list",
-  "CIVIL_AND_FAMILY_DAILY_CAUSE_LIST": "civil-and-family-daily-cause-list"
+  "CIVIL_AND_FAMILY_DAILY_CAUSE_LIST": "civil-and-family-daily-cause-list",
+  "SSCS_DAILY_LIST": "sscs-daily-list"
 }
