+++ src/main/resources/locales/bill/sscs-daily-list.json
@@ -1,0 +1,23 @@
+{
+  "title": "SSCS Daily List",
+  "hearingsFor": " hearings for ",
+  "hearingsForBill": " gwrandawiadau ar gyfer ",
+  "listUpdated": "Last Updated/Diweddarwyd diwethaf DATE at/yn ",
+  "listDate": "List for/Rhestr ar gyfer",
+  "importantInformationHeading": "Important information/Gwybodaeth Bwysig",
+  "importantInformationP1": "Open justice is a fundamental principle in our courts and tribunals system, and will continue to be as we develop the use of audio and video technology.",
+  "importantInformationP2": "Requests from the media and others to observe a hearing remotely in NAME should be made direct to: EMAIL or by calling TELNO. Arrangements will then be made to enable you to attend.",
+  "importantInformationP3": "When considering the use of telephone and video technology, the judiciary will have regard to the principles of open justice. Judges may determine that a hearing should be held in private if this is necessary to secure the proper administration of justice.",
+  "importantInformationP4": "Mae mynediad at gyfiawnder yn egwyddor hanfodol yn system y llysoedd a’r tribiwnlysoedd, a bydd hyn yn parhau wrth i ni ddatblygu technoleg sain a fideo.",
+  "importantInformationP5": "Dylid cyfeirio ceisiadau gan y cyfryngau ac eraill i arsylwi gwrandawiad o bell yn NAME at: EMAIL neu drwy ffonio TELNO. Yna fe wneir trefniadau i’ch galluogi i fynychu.",
+  "importantInformationP6": "Wrth ystyried a ddylid defnyddio technoleg ffôn a fideo, bydd y farnwriaeth yn ystyried yr egwyddor mynediad at gyfiawnder. Efallai y bydd barnwyr yn penderfynu y dylid cynnal gwrandawiad yn breifat os yw hyn yn angenrheidiol i sicrhau y gweinyddir cyfiawnder yn briodol.",
+  "dataSource": "Data Source/Fynhonnell Data",
+  "backButton": "Back/Yn ôl",
+  "hearingTime": "Hearing Time/Amser y Gwrandawiad",
+  "appRefNumber": "Appeal reference number/Cyfeirnod Apêl",
+  "tribunalType": "Tribunal type/Math o Dribiwnlys",
+  "appellant": "Appellant/Apelydd",
+  "respondent": "FTA/Respondent/FTA/Atebydd",
+  "panel": "Panel",
+  "legalAdvisor": "Legal Advisor/Cynghorydd Cyfreithiol"
+}
