+++ src/main/resources/locales/en/sscs-daily-list.json
@@ -1,0 +1,19 @@
+{
+  "title": "SSCS Daily List",
+  "hearingsFor": " hearings for ",
+  "listUpdated": "Last Updated DATE at ",
+  "listDate": "List for",
+  "importantInformationHeading": "Important information",
+  "importantInformationP1": "Open justice is a fundamental principle in our courts and tribunals system, and will continue to be as we develop the use of audio and video technology.",
+  "importantInformationP2": "Requests from the media and others to observe a hearing remotely in NAME should be made direct to: EMAIL or by calling TELNO. Arrangements will then be made to enable you to attend.",
+  "importantInformationP3": "When considering the use of telephone and video technology, the judiciary will have regard to the principles of open justice. Judges may determine that a hearing should be held in private if this is necessary to secure the proper administration of justice.",
+  "dataSource": "Data Source",
+  "backButton": "Back",
+  "hearingTime": "Hearing Time",
+  "appRefNumber": "Appeal reference number",
+  "tribunalType": "Tribunal type",
+  "appellant": "Appellant",
+  "respondent": "FTA/Respondent",
+  "panel": "Panel",
+  "legalAdvisor": "Legal Advisor"
+}
