+++ src/main/views/sscs-daily-list.njk
@@ -1,0 +1,106 @@
+{% from "govuk/components/table/macro.njk" import govukTable as 'govukTable' %}
+{% from "./macros/common-components.njk" import goBack %}
+
+{% extends "template.njk" %}
+{% block pageTitle %}
+  {{ title }}
+{% endblock %}
+{% block beforeContent %}
+  {{ super() }}
+  {{ goBack(text = backButton) }}
+{% endblock %}
+{% block content %}
+<div class="parent-box">
+  <h2 id="page-heading" class="govuk-heading-l">
+    {{ courtName }}{{ hearingsFor }}{{ contentDate }}
+    {% if bill %}
+      /{{ courtName }}{{ hearingsForBill }}{{ contentDate }}
+    {% endif %}
+  </h2>
+  <p class="govuk-body govuk-!-font-weight-bold govuk-!-margin-bottom-1">{{ listDate }} {{ contentDate }}</p>
+  <p class="govuk-body">{{ listUpdated | replace("DATE", publishedDate) }} {{ publishedTime }}</p>
+  <details class="govuk-details" data-module="govuk-details" open>
+    <summary class="govuk-details__summary"><span class="govuk-details__summary-text">{{ importantInformationHeading }}</span></summary>
+    <div class="govuk-details__text">
+      <p>{{ importantInformationP1 }}</p>
+      <p>{{ importantInformationP2 | replace("NAME", courtName) | replace("EMAIL", listData['venue']["venueContact"]["venueEmail"]) | replace("TELNO", listData['venue']['venueContact']['venueTelephone'])}}</p>
+      <p>{{ importantInformationP3 }}</p>
+      {% if bill %}
+        <P>{{ importantInformationP4 }}</P>
+        <P>{{ importantInformationP5 | replace("NAME", courtName) | replace("EMAIL", listData['venue']["venueContact"]["venueEmail"]) | replace("TELNO", listData['venue']['venueContact']['venueTelephone'])}}</P>
+        <P>{{ importantInformationP6 }}</P>
+      {% endif %}
+    </div>
+  </details>
+  {% for courtList in listData['courtLists'] %}
+    {% set courtHouse = courtList['courtHouse'] %}
+    <div class="site-header" tabindex="0" aria-label="Court Details">
+      <h1 class="govuk-heading-l site-address" tabindex="0">{{ courtHouse['courtHouseName'] }}</h1>
+    </div>
+
+    <div class="govuk-accordion" data-module="govuk-accordion" id="accordion-default">
+      {% set courtRoomCount = 0 %}
+      {% for courtRoom in courtHouse['courtRoom'] %}
+        {% set hearingCount = 1 %}
+        {% set courtRoomCount = courtRoomCount + 1 %}
+      <div class="govuk-accordion__section govuk-accordion__section--expanded">
+        <div class="govuk-accordion__section-header">
+          <h2 class="govuk-accordion__section-heading">
+            <span class="govuk-accordion__section-button" id="accordion-default-heading-{{ courtRoomCount }}">
+              {{ courtRoom['courtRoomName'] }}
+            </span>
+          </h2>
+        </div>
+        <div id="accordion-default-content-{{ courtRoomCount }}" class="govuk-accordion__section-content" aria-labelledby="accordion-default-heading-{{ courtRoomCount }}">
+          <div class="parent-box overflow-table">
+            <table class="govuk-table">
+              <thead class="govuk-table__head">
+              <tr class="govuk-table__row">
+                <th scope="col" class="govuk-table__header">{{ hearingTime }}</th>
+                <th scope="col" class="govuk-table__header">{{ appRefNumber }}</th>
+                <th scope="col" class="govuk-table__header">{{ tribunalType }}</th>
+                <th scope="col" class="govuk-table__header">{{ appellant }}</th>
+                <th scope="col" class="govuk-table__header">{{ respondent }}</th>
+                <th scope="col" class="govuk-table__header">{{ panel }}</th>
+              </tr>
+              </thead>
+              <tbody class="govuk-table__body">
+              {% for session in courtRoom['session'] %}
+                {% for sitting in session['sittings'] %}
+                  {% for hearing in sitting['hearing'] %}
+                    {% for case in hearing['case'] %}
+                      {% if courtRoom['totalHearing'] === hearingCount %}
+                        {% set noBorder = 'no-border-bottom' %}
+                      {% endif %}
+                      <tr class="govuk-table__row">
+                        <td class="govuk-table__cell {{ noBorder }}">{{ sitting['sittingStartFormatted'] }}</td>
+                        <td class="govuk-table__cell {{ noBorder }}">{{ case['caseNumber'] }}</td>
+                        <td class="govuk-table__cell {{ noBorder }}">{{ sitting['caseHearingChannel'] }}</td>
+                        <td class="govuk-table__cell {{ noBorder }}">{{ hearing['applicant'] | replace("LEGALADVISOR", legalAdvisor) }}</td>
+                        <td class="govuk-table__cell {{ noBorder }}">{{ hearing['prosecutionAuthorityRefFormatted'] }}</td>
+                        <td class="govuk-table__cell {{ noBorder }}">{{ session['formattedJudiciary'] }}</td>
+                      </tr>
+                    {% endfor %}
+                  {% endfor %}
+                {% endfor %}
+              {% endfor %}
+              </tbody>
+            </table>
+          </div>
+        </div>
+      </div>
+      {% endfor %}
+    </div>
+  {% endfor %}
+  <p class="govuk-body govuk-!-font-size-14">{{ dataSource }}: {{ provenance }}</p>
+  {{ super() }}
+</div>
+{% endblock %}
+{% block bodyEnd %}
+  <script>
+    if ( window.history.replaceState ) {
+      window.history.replaceState( null, null, window.location.href );
+    }
+  </script>
+  {{ super() }}
+{% endblock %}
