+++ src/test/routes/family-daily-cause-list.test.ts
@@ -6,12 +6,13 @@ import { app } from '../../main/app';
 import { PublicationService } from '../../main/service/publicationService';
 import fs from 'fs';
 import path from 'path';
+import { DataManipulationService } from '../../main/service/dataManipulationService';
 
 const rawData = fs.readFileSync(path.resolve(__dirname, '../unit/mocks/familyDailyCauseList.json'), 'utf-8');
 const dailyReferenceData = JSON.parse(rawData);
 sinon.stub(PublicationService.prototype, 'getIndividualPublicationJson').resolves(dailyReferenceData);
 sinon.stub(PublicationService.prototype, 'getIndividualPublicationMetadata').resolves(dailyReferenceData);
-sinon.stub(PublicationService.prototype, 'manipulatedDailyListData').resolves(dailyReferenceData);
+sinon.stub(DataManipulationService.prototype, 'manipulatedDailyListData').resolves(dailyReferenceData);
 
 describe('Daily Cause List Page', () => {
   describe('on GET', () => {
