+++ src/test/unit/controllers/DailyCauseListController.test.ts
@@ -7,6 +7,7 @@ import { PublicationService } from '../../../main/service/publicationService';
 import {mockRequest} from '../mocks/mockRequest';
 import moment from 'moment';
 import {LocationService} from '../../../main/service/locationService';
+import {DataManipulationService} from '../../../main/service/dataManipulationService';
 
 const rawData = fs.readFileSync(path.resolve(__dirname, '../mocks/familyDailyCauseList.json'), 'utf-8');
 const listData = JSON.parse(rawData);
@@ -22,7 +23,7 @@ const dailyCauseListController = new DailyCauseListController();
 const dailyCauseListJsonStub = sinon.stub(PublicationService.prototype, 'getIndividualPublicationJson');
 const dailyCauseListMetaDataStub = sinon.stub(PublicationService.prototype, 'getIndividualPublicationMetadata');
 sinon.stub(LocationService.prototype, 'getLocationById').resolves(courtData[0]);
-sinon.stub(PublicationService.prototype, 'manipulatedDailyListData').returns(listData);
+sinon.stub(DataManipulationService.prototype, 'manipulatedDailyListData').returns(listData);
 
 const artefactId = 'abc';
