+++ src/test/unit/service/manualUploadService.test.ts
@@ -76,7 +76,7 @@ describe('Manual upload service', () => {
 
     it('should build form data list subtypes', async () => {
       const data = await manualUploadService.buildFormData(englishLanguage);
-      expect(data['listSubtypes'].length).to.equal(10);
+      expect(data['listSubtypes'].length).to.equal(11);
       expect(data['listSubtypes'][0]).to.deep.equal({text:'SJP Public List', value: 'SJP_PUBLIC_LIST'});
     });
