+++ src/test/unit/views/sscs-daily-list-test.ts
@@ -1,0 +1,63 @@
+import fs from 'fs';
+import path from 'path';
+import {PublicationService} from '../../../main/service/publicationService';
+import {LocationService} from '../../../main/service/locationService';
+import sinon from 'sinon';
+import request from 'supertest';
+import {app} from '../../../main/app';
+import {expect} from 'chai';
+
+const PAGE_URL = '/sscs-daily-list?artefactId=abc';
+const headingClass = 'govuk-heading-l';
+const summaryHeading = 'govuk-details__summary-text';
+const summaryText = 'govuk-details__text';
+
+const courtName = 'Abergavenny Magistrates\' Court';
+const expectedHeader = courtName + ' hearings for';
+const summaryHeadingText = 'Important information';
+let htmlRes: Document;
+
+const rawData = fs.readFileSync(path.resolve(__dirname, '../mocks/sscsDailyList.json'), 'utf-8');
+const sscsDailyList = JSON.parse(rawData);
+const rawMetaData = fs.readFileSync(path.resolve(__dirname, '../mocks/returnedArtefacts.json'), 'utf-8');
+const metaData = JSON.parse(rawMetaData)[0];
+
+const rawDataCourt = fs.readFileSync(path.resolve(__dirname, '../mocks/courtAndHearings.json'), 'utf-8');
+const courtData = JSON.parse(rawDataCourt);
+
+sinon.stub(PublicationService.prototype, 'getIndividualPublicationJson').returns(sscsDailyList);
+sinon.stub(PublicationService.prototype, 'getIndividualPublicationMetadata').returns(metaData);
+sinon.stub(LocationService.prototype, 'getLocationById').resolves(courtData[0]);
+
+describe('Sscs daily list page', () => {
+  beforeAll(async () => {
+    await request(app).get(PAGE_URL).then(res => {
+      htmlRes = new DOMParser().parseFromString(res.text, 'text/html');
+    });
+  });
+
+  it('should display header',  () => {
+    const header = htmlRes.getElementsByClassName(headingClass);
+    expect(header[0].innerHTML).contains(expectedHeader, 'Could not find the header');
+  });
+
+  it('should display summary',  () => {
+    const summary = htmlRes.getElementsByClassName(summaryHeading);
+    expect(summary[0].innerHTML).contains(summaryHeadingText, 'Could not find the display summary heading');
+  });
+
+  it('should display court name summary paragraph',  () => {
+    const summary = htmlRes.getElementsByClassName(summaryText);
+    expect(summary[0].innerHTML).contains(courtName, 'Could not find the court name in summary text');
+  });
+
+  it('should display court email summary paragraph',  () => {
+    const summary = htmlRes.getElementsByClassName(summaryText);
+    expect(summary[0].innerHTML).contains('court1@moj.gov.u', 'Could not find the court name in summary text');
+  });
+
+  it('should display court contact number summary paragraph',  () => {
+    const summary = htmlRes.getElementsByClassName(summaryText);
+    expect(summary[0].innerHTML).contains('01772 844700', 'Could not find the court name in summary text');
+  });
+});
